requirements.txt
-i https://pypi.org/simple
alembic==1.13.3; python_version >= '3.8'
aniso8601==9.0.1
asttokens==2.4.1
attrs==24.2.0; python_version >= '3.7'
backcall==0.2.0
bcrypt==4.2.0; python_version >= '3.7'
click==8.1.7; python_version >= '3.7'
decorator==5.1.1; python_version >= '3.5'
exceptiongroup==1.2.2; python_version >= '3.7'
executing==2.1.0; python_version >= '3.8'
faker==15.3.2; python_version >= '3.7'
flask==2.2.5; python_version >= '3.7'
flask-bcrypt==1.0.1
flask-cors==5.0.0
flask-migrate==4.0.7; python_version >= '3.6'
flask-restful==0.3.10
flask-sqlalchemy==3.1.1; python_version >= '3.8'
greenlet==3.1.1; python_version >= '3.7'
gunicorn==23.0.0; python_version >= '3.7'
importlib-metadata==6.0.0; python_version >= '3.7'
importlib-resources==5.10.0; python_version >= '3.7'
iniconfig==2.0.0; python_version >= '3.7'
ipdb==0.13.9; python_version >= '2.7'
ipython==8.12.3; python_version >= '3.8'
itsdangerous==2.2.0; python_version >= '3.8'
jedi==0.19.1; python_version >= '3.6'
jinja2==3.1.4; python_version >= '3.7'
mako==1.3.5; python_version >= '3.8'
markupsafe==2.1.5; python_version >= '3.7'
matplotlib-inline==0.1.7; python_version >= '3.8'
packaging==24.1; python_version >= '3.8'
parso==0.8.4; python_version >= '3.6'
pexpect==4.9.0
pickleshare==0.7.5
pluggy==1.5.0; python_version >= '3.8'
prompt-toolkit==3.0.48; python_full_version >= '3.7.0'
psycopg2-binary==2.9.9; python_version >= '3.7'
ptyprocess==0.7.0
pure-eval==0.2.3
pygments==2.18.0; python_version >= '3.8'
pytest==7.2.0; python_version >= '3.7'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
python-dotenv==0.21.1
pytz==2024.2
setuptools==75.1.0; python_version >= '3.8'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
sqlalchemy==2.0.35; python_version >= '3.7'
sqlalchemy-serializer==1.4.12
stack-data==0.6.3
toml==0.10.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2'
tomli==2.0.2; python_version >= '3.8'
traitlets==5.14.3; python_version >= '3.8'
typing-extensions==4.12.2; python_version >= '3.8'
wcwidth==0.2.13
werkzeug==2.2.2; python_version >= '3.7'
zipp==3.20.2; python_version >= '3.8'
